refactor: PointGenerator
The paper specified a Point Generator, that is equivalent to the
PCT_refine that was in the code. In this commit I renamed that to Point
Generator to be in line with what the paper said (so it's easier to
follow the code).

There was also an unused variable in the forward pass of the PCT_refine,
so here it was removed.