feat: Changes de SG so it predicts only the break
I tried to see what happened if the Seed Generator tried to predict how
was the missing piece (restoration shape), instead of predicting the
whole object. So, part of the loss now comes from computing the CD
between the coarse (points before the concatenation in the Seed
Generator) and the restoration shape, since in our dataset inspired by
the Geometric Breaks Dataset we have that information.

This actually worked kind of well, but the Point Generator now sort of
ruins the object.