refactor: Changes cd value when net gets 0 points.
I was setting an infinity cd value when there were no points in the
filtered shape, ensuring the net to decide points. This was changed to
avoid the net be too punished, because this made the rest of the loss
irrelevant and I feared the net wouldn't be able to learn that much (It
learned, since the experiments worked, but I think it can learn better).

This change made the net to take twice long to complete all the epochs
tho.